reached LeafDecided event in test_upgrade_task
